Add files via upload
This Arduino project serves as a basic introduction to programming with the microcontroller. Commonly referred to as the "Hello World" project, it's an ideal starting point for beginners. The program simply configures the Arduino to display or print the message "Hello, World!" It's a straightforward way to get acquainted with Arduino programming.